- move raft impl to raft.go - remove Ready.RetransmitMsgApps()
