+++ pkg/kv/kvserver/kvflowcontrol/kvflowconnectedstream/BUILD.bazel
@@ -3,6 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "kvflowconnectedstream",
     srcs = [
+        "raft.go",
         "store_stream.go",
         "stream.go",
         "token_counter.go",
